Load config on-demand, config.NotSet, config debug
Instead of importing `config.config`, use `config.config_get`.
If `config` hasn't be loaded yet, `config_get` will load it and
also print debug info about the caller that triggered the config
load.

The `NotSet` object was moved to config, where it is used, and can
easily be imported for use in other modules (`test_tigertamer.py`).

Config is loaded way before args are parsed, so any `debug()` calls
from the `config` module were unconditional.
`-D` and `--debug` are checked in `config` to enable/disable the
debug printer. This way, the debug flags are part of the configuration
if TigerTamer, and config can use debug like any other module.