Move `show_error` to `WinToplevelBase`.
Any `Toplevel` window needs to get out of the way
when error dialogs are shown. Especially if they are
set to `-topmost=1`. The code to gracefully show an
error dialog (and do the right thing) was moved into
the base class. Now, no matter what the `-topmost`
setting is on the window, `show_error` will temporarily
remove the `-topmost` setting, show the error, and
then restore it back to whatever it was before.